backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
email-validator==2.2.0
pandas==2.3.2
openpyxl==3.1.5
python-dateutil==2.9.0.post0
slowapi==0.1.9
httpx==0.27.2
orjson==3.10.12

pytest==8.4.1
pytest-asyncio==0.23.8
pytest-cov==5.0.0
aiosqlite==0.20.0
